services/backend/migrations/models/1_20240123161904_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "entry" ADD "supplier" VARCHAR(225) NOT NULL;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "entry" DROP COLUMN "supplier";"""
