services/backend/pyproject.toml
[tool.poetry]
name = "ledger"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "*"
fastapi = "*"
tortoise-orm = "*"
aerich = "*"
asyncpg = "*"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
mypy = "*"
pip-audit = "*"
ruff = "*"
pytest = "*"
pytest-cov = "*"
httpx = "*"


[tool.aerich]
tortoise_orm = "src.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
